refactor: Add common.py to hold logger and tqdm wrapper, Disable tqdm progress on verbose mode, Remove Unused imports
